Add mutation testing (Closes #13)
This commit adds PIT as the mutation test framework for all subprojects within this repository.

Later on, the report will be calculated ones a week by a GitHub cronjob and updated in the UI. However, for now the report can be called only manually as the calculation time of it takes some time.