Add Dockerfile (Closes #6)
This commit adds Dockerfile to the three services in this project. While `gateway` and `data-processor` are not producing any artifact at this point, the foundations for the Dockerfile have already been set so CI can build them successfully.

Moreover, `client-app` was removed from the current CI as it requires a different set of CI pipelines to work in comparison to the backend services.