Introduce SpringBoot Gateway (Closes #40) (#63)
This commit introduces the minimum required setup to run the Gateway service using SpringBoot and the hexagonal architecture.  Moreover, it fixes the jar and dockerfile of the services to make it properly work.

While the service has no meaningful code at this point (it only starts), it contains all the required components to add them later on.