Setup Debezium (Closes #41)
This commit creates a simple integration between the fetcher service and debezium allowing implementation of the outbox pattern for fetched comics.

The current setup requires manual creation of the connector using the http collection committed to the repo, but this should be automated in the future to be part of the docker compose startup process.