Build only submodule
This commit ensures that each submodule in the project (AKA deployable unit / service) is built by itself, and not the entire project as a whole.

This will decrease build time as different components can be built in parallel, and components that have not changed will not be built at all.